Restructure (de)serialization code
