+++ integration-testing/integration-tests.cabal
@@ -30,8 +30,6 @@ executable integration-tests
     , base >=4.7 && <5
     , bytestring
 
-    -- External
-    , http-client < 0.8
   other-modules:
       Paths_integration_tests
       IntegrationTests.QuerySerialization
